+++ Tests/GravatarTests/AvatarServiceTests.swift
@@ -129,13 +129,3 @@ private func avatarService(with session: URLSessionProtocol, cache: ImageCaching
     let service = AvatarService(client: client, cache: cache)
     return service
 }
-
-extension HTTPURLResponse {
-    static func successResponse(with url: URL? = URL(string: "https://gravatar.com")) -> HTTPURLResponse {
-        HTTPURLResponse(url: url!, statusCode: 200, httpVersion: nil, headerFields: nil)!
-    }
-
-    static func errorResponse(with url: URL? = URL(string: "https://gravatar.com"), code: Int) -> HTTPURLResponse {
-        HTTPURLResponse(url: url!, statusCode: code, httpVersion: nil, headerFields: nil)!
-    }
-}
