+++ Tests/GravatarTests/ImageDownloadServiceTests.swift
@@ -1,5 +1,4 @@
 @testable import Gravatar
-@testable import GravatarUI
 import TestHelpers
 import XCTest
 
@@ -14,27 +13,6 @@ final class ImageDownloadServiceTests: XCTestCase {
 
         XCTAssertEqual(sessionMock.request?.url?.absoluteString, "https://gravatar.com/avatar/HASH")
         XCTAssertNotNil(imageResponse.image)
-    }
-
-    func testFetchImageWithCompletionHandlerAndURL() {
-        let imageURL = "https://gravatar.com/avatar/HASH"
-        let response = HTTPURLResponse.successResponse(with: URL(string: imageURL))
-        let sessionMock = URLSessionMock(returnData: ImageHelper.testImageData, response: response)
-        let service = imageDownloadService(with: sessionMock)
-        let expectation = expectation(description: "Request finishes")
-
-        service.fetchImage(with: URL(string: imageURL)!) { response in
-            switch response {
-            case .success(let result):
-                XCTAssertNotNil(result.image)
-                XCTAssertEqual(result.sourceURL.absoluteString, imageURL)
-            case .failure(let error):
-                XCTFail(error.localizedDescription)
-            }
-            expectation.fulfill()
-        }
-
-        wait(for: [expectation], timeout: 0.2)
     }
 
     func testFetchCatchedImageWithURL() async throws {
@@ -53,26 +31,6 @@ final class ImageDownloadServiceTests: XCTestCase {
         XCTAssertEqual(sessionMock.callsCount, 1)
         XCTAssertEqual(sessionMock.request?.url?.absoluteString, "https://gravatar.com/avatar/HASH")
         XCTAssertNotNil(imageResponse.image)
-    }
-
-    func testFetchImageWithCompletionHandlerError() throws {
-        let imageURL = try XCTUnwrap(URL(string: "https://gravatar.com/avatar/HASH"))
-        let response = HTTPURLResponse.errorResponse(code: 404)
-        let sessionMock = URLSessionMock(returnData: ImageHelper.testImageData, response: response)
-        let service = imageDownloadService(with: sessionMock)
-        let expectation = expectation(description: "Request finishes")
-
-        service.fetchImage(with: imageURL) { response in
-            switch response {
-            case .failure(.responseError(reason: let reason)) where reason.httpStatusCode == 404:
-                break
-            default:
-                XCTFail("Request should fail")
-            }
-            expectation.fulfill()
-        }
-
-        wait(for: [expectation], timeout: 0.2)
     }
 }
